Updated Files
I changed the links in the index.html files form "login.html" to /login since that is what we ref for in the main code. Also did the same thing for sign up. Also added spam detection in the signup page its and invisible text field that the user cannot see but it is still writable. This detects a spam bot because when this field is filled it would mean that a person is not the one filling out the questions. Also I added the homepage and the links for the profiles page and settings page. Finally I added the SendEmailmain() method that in the event the user does not enter their information correctly an email gets sent out.  
Still need to do: compare the locations/device the user has registered in the database and send the alert out based on that currently the default message is sending out a location alert email

Files touched: index, main, homepage